Wait for dns to be configured
A prior fix for bug 1838607 removed the wait
for networking to be configured which has had
the knock-on effect that if vaultlocker is
using hostnames instead of ipaddresses it will
not be able to resolve them if dns is not yet
configured. This patch adds After=nss-lookup.target
to ensure dns is configured prior to starting.

Closes-Bug: 1868557